+++ scripts/tests/calibnet_export_check.sh
@@ -3,14 +3,14 @@
 # the snapshot export feature.
 # It requires both the `forest` and `forest-cli` binaries to be in the PATH.
 
-set -e
+set -eu
 
 source "$(dirname "$0")/harness.sh"
 
 forest_init
 
 echo "Cleaning up the initial snapshot"
-rm -rf ./*.car.*
+rm --force --verbose ./*.{car,car.zst,sha256sum}
 
 echo "Exporting zstd compressed snapshot"
 $FOREST_CLI_PATH snapshot export
@@ -33,3 +33,10 @@ pushd "$(mktemp --directory)"
     "$FOREST_PATH" --chain calibnet --encrypt-keystore false --halt-after-import --no-gc --import-snapshot "$diffed_snapshot"
 rm -- *
 popd
+
+echo "Validating CAR files"
+zstd --decompress ./*.car.zst
+for f in *.car; do
+  echo "Validating CAR file $f"
+  $FOREST_CLI_PATH --chain calibnet snapshot validate "$f"
+done