+++ src/blocks/tipset.rs
@@ -1,12 +1,13 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use std::fmt;
+use std::{fmt, sync::OnceLock};
 
-use crate::shim::address::Address;
-use crate::shim::clock::ChainEpoch;
+use crate::networks::{calibnet, mainnet};
+use crate::shim::{address::Address, clock::ChainEpoch};
 use crate::utils::cid::CidCborExt;
-use ahash::{HashSet, HashSetExt};
+use ahash::{HashMap, HashSet, HashSetExt};
+use anyhow::Context as _;
 use cid::Cid;
 use fvm_ipld_blockstore::Blockstore;
 use fvm_ipld_encoding::CborStore;
@@ -100,40 +101,6 @@ impl quickcheck::Arbitrary for Tipset {
     }
 }
 
-#[cfg(test)]
-mod property_tests {
-    use quickcheck_macros::quickcheck;
-    use serde_json;
-
-    use super::{
-        tipset_json::{TipsetJson, TipsetJsonRef},
-        tipset_keys_json::TipsetKeysJson,
-        Tipset, TipsetKeys,
-    };
-
-    impl quickcheck::Arbitrary for TipsetKeys {
-        fn arbitrary(g: &mut quickcheck::Gen) -> Self {
-            Self {
-                cids: Vec::arbitrary(g),
-            }
-        }
-    }
-
-    #[quickcheck]
-    fn tipset_keys_roundtrip(tipset_keys: TipsetKeys) {
-        let serialized = serde_json::to_string(&TipsetKeysJson(tipset_keys.clone())).unwrap();
-        let parsed: TipsetKeysJson = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(tipset_keys, parsed.0);
-    }
-
-    #[quickcheck]
-    fn tipset_roundtrip(tipset: Tipset) {
-        let serialized = serde_json::to_string(&TipsetJsonRef(&tipset)).unwrap();
-        let parsed: TipsetJson = serde_json::from_str(&serialized).unwrap();
-        assert_eq!(&tipset, parsed.0.as_ref());
-    }
-}
-
 impl From<FullTipset> for Tipset {
     fn from(full_tipset: FullTipset) -> Self {
         let key = full_tipset.key;
@@ -169,7 +136,8 @@ impl Tipset {
         })
     }
 
-    /// Loads a tipset from memory given the tipset keys.
+    /// Fetch a tipset from the blockstore. This call fails if the tipset
+    /// present but invalid. If the tipset is missing, None is returned.
     pub fn load(store: impl Blockstore, tsk: &TipsetKeys) -> anyhow::Result<Option<Tipset>> {
         Ok(tsk
             .cids()
@@ -178,6 +146,12 @@ impl Tipset {
             .collect::<anyhow::Result<Option<_>>>()?
             .map(Tipset::new)
             .transpose()?)
+    }
+
+    /// Fetch a tipset from the blockstore. This calls fails if the tipset is
+    /// missing or invalid.
+    pub fn load_required(store: impl Blockstore, tsk: &TipsetKeys) -> anyhow::Result<Tipset> {
+        Tipset::load(store, tsk)?.context("Required tipset missing from database")
     }
 
     /// Constructs and returns a full tipset if messages from storage exists
@@ -290,6 +264,46 @@ impl Tipset {
                 child
             })
         })
+    }
+
+    /// Fetch the genesis block header for a given tipset.
+    pub fn genesis(&self, store: impl Blockstore) -> anyhow::Result<BlockHeader> {
+        // Scanning through millions of epochs to find the genesis is quite
+        // slow. Let's use a list of known blocks to short-circuit the search.
+        // The blocks are hash-chained together and known blocks are guaranteed
+        // to have a known genesis.
+        #[derive(Serialize, Deserialize)]
+        struct KnownHeaders {
+            calibnet: HashMap<ChainEpoch, String>,
+            mainnet: HashMap<ChainEpoch, String>,
+        }
+
+        static KNOWN_HEADERS: OnceLock<KnownHeaders> = OnceLock::new();
+        let headers = KNOWN_HEADERS.get_or_init(|| {
+            serde_yaml::from_str(include_str!("../../build/known_blocks.yaml")).unwrap()
+        });
+
+        for tipset in self.clone().chain(&store) {
+            // Search for known calibnet and mainnet blocks
+            for (genesis_cid, known_blocks) in [
+                (*calibnet::GENESIS_CID, &headers.calibnet),
+                (*mainnet::GENESIS_CID, &headers.mainnet),
+            ] {
+                if let Some(known_block_cid) = known_blocks.get(&tipset.epoch()) {
+                    if known_block_cid == &tipset.min_ticket_block().cid().to_string() {
+                        return store
+                            .get_cbor(&genesis_cid)?
+                            .ok_or_else(|| anyhow::anyhow!("Genesis block missing from database"));
+                    }
+                }
+            }
+
+            // If no known blocks are found, we'll eventually hit the genesis tipset.
+            if tipset.epoch() == 0 {
+                return Ok(tipset.min_ticket_block().clone());
+            }
+        }
+        anyhow::bail!("Genesis block not found")
     }
 }
 
@@ -516,6 +530,40 @@ pub mod tipset_json {
         }
         let TipsetDe { blocks, .. } = Deserialize::deserialize(deserializer)?;
         Tipset::new(blocks).map(Arc::new).map_err(de::Error::custom)
+    }
+}
+
+#[cfg(test)]
+mod property_tests {
+    use quickcheck_macros::quickcheck;
+    use serde_json;
+
+    use super::{
+        tipset_json::{TipsetJson, TipsetJsonRef},
+        tipset_keys_json::TipsetKeysJson,
+        Tipset, TipsetKeys,
+    };
+
+    impl quickcheck::Arbitrary for TipsetKeys {
+        fn arbitrary(g: &mut quickcheck::Gen) -> Self {
+            Self {
+                cids: Vec::arbitrary(g),
+            }
+        }
+    }
+
+    #[quickcheck]
+    fn tipset_keys_roundtrip(tipset_keys: TipsetKeys) {
+        let serialized = serde_json::to_string(&TipsetKeysJson(tipset_keys.clone())).unwrap();
+        let parsed: TipsetKeysJson = serde_json::from_str(&serialized).unwrap();
+        assert_eq!(tipset_keys, parsed.0);
+    }
+
+    #[quickcheck]
+    fn tipset_roundtrip(tipset: Tipset) {
+        let serialized = serde_json::to_string(&TipsetJsonRef(&tipset)).unwrap();
+        let parsed: TipsetJson = serde_json::from_str(&serialized).unwrap();
+        assert_eq!(&tipset, parsed.0.as_ref());
     }
 }