+++ src/daemon/mod.rs
@@ -171,7 +171,10 @@ pub(super) async fn start(
     let keystore = Arc::new(RwLock::new(keystore));
 
     let chain_data_path = chain_path(&config);
-    let db = open_proxy_db(db_root(&chain_data_path), config.db_config().clone())?;
+    let db = Arc::new(open_proxy_db(
+        db_root(&chain_data_path),
+        config.db_config().clone(),
+    )?);
 
     let mut services = JoinSet::new();
 
@@ -213,7 +216,7 @@ pub(super) async fn start(
 
     // Initialize ChainStore
     let chain_store = Arc::new(ChainStore::new(
-        db.clone(),
+        Arc::clone(&db),
         config.chain.clone(),
         &genesis_header,
         chain_data_path.as_path(),
@@ -226,7 +229,7 @@ pub(super) async fn start(
         let chain_store = chain_store.clone();
         let get_tipset = move || chain_store.heaviest_tipset().as_ref().clone();
         Arc::new(DbGarbageCollector::new(
-            db,
+            db.as_ref().clone(),
             file_backed_chain_meta,
             config.chain.policy.chain_finality,
             config.chain.recent_state_roots,
@@ -247,18 +250,8 @@ pub(super) async fn start(
 
     let publisher = chain_store.publisher();
 
-    // Reward calculation is needed by the VM to calculate state, which can happen
-    // essentially anywhere the `StateManager` is called. It is consensus
-    // specific, but threading it through the type system would be a nightmare,
-    // which is why dynamic dispatch is used.
-    let reward_calc = cns::reward_calc();
-
     // Initialize StateManager
-    let sm = StateManager::new(
-        Arc::clone(&chain_store),
-        Arc::clone(&config.chain),
-        reward_calc,
-    )?;
+    let sm = StateManager::new(Arc::clone(&chain_store), Arc::clone(&config.chain))?;
 
     let state_manager = Arc::new(sm);
 
@@ -306,7 +299,7 @@ pub(super) async fn start(
         net_keypair,
         &network_name,
         genesis_cid,
-    )?;
+    );
 
     let network_rx = p2p_service.network_receiver();
     let network_send = p2p_service.network_sender();
@@ -317,8 +310,8 @@ pub(super) async fn start(
         provider,
         network_name.clone(),
         network_send.clone(),
-        MpoolConfig::load_config(&db)?,
-        Arc::clone(state_manager.chain_config()),
+        MpoolConfig::load_config(db.as_ref())?,
+        state_manager.chain_config(),
         &mut services,
     )?;
 
@@ -763,7 +756,7 @@ mod test {
     }
 
     async fn import_snapshot_from_file(file_path: &str) -> anyhow::Result<()> {
-        let db = MemoryDB::default();
+        let db = Arc::new(MemoryDB::default());
         let chain_config = Arc::new(ChainConfig::default());
 
         let genesis_header = BlockHeader::builder()
@@ -778,11 +771,7 @@ mod test {
             &genesis_header,
             chain_data_root.path(),
         )?);
-        let sm = Arc::new(StateManager::new(
-            cs,
-            chain_config,
-            Arc::new(crate::interpreter::RewardActorMessageCalc),
-        )?);
+        let sm = Arc::new(StateManager::new(cs, chain_config)?);
         import_chain::<_>(
             &sm,
             file_path,
@@ -796,7 +785,7 @@ mod test {
 
     #[tokio::test]
     async fn import_chain_from_file() -> anyhow::Result<()> {
-        let db = MemoryDB::default();
+        let db = Arc::new(MemoryDB::default());
         let chain_config = Arc::new(ChainConfig::default());
         let genesis_header = BlockHeader::builder()
             .miner_address(Address::new_id(0))
@@ -810,11 +799,7 @@ mod test {
             &genesis_header,
             chain_data_root.path(),
         )?);
-        let sm = Arc::new(StateManager::new(
-            cs,
-            chain_config,
-            Arc::new(crate::interpreter::RewardActorMessageCalc),
-        )?);
+        let sm = Arc::new(StateManager::new(cs, chain_config)?);
         import_chain::<_>(
             &sm,
             "test-snapshots/chain4.car",