+++ src/interpreter/fvm3.rs
@@ -5,6 +5,8 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::{cell::Ref, sync::Arc};
 
 use crate::blocks::BlockHeader;
+use crate::blocks::Tipset;
+use crate::chain::store::ChainStore;
 use crate::interpreter::errors::Error;
 use crate::networks::ChainConfig;
 use crate::shim::{
@@ -32,35 +34,41 @@ use crate::interpreter::resolve_to_key_addr;
 
 pub struct ForestExterns<DB> {
     rand: Box<dyn Rand>,
+    heaviest_tipset: Arc<Tipset>,
     epoch: ChainEpoch,
     root: Cid,
-    lookback: Box<dyn Fn(ChainEpoch) -> anyhow::Result<Cid>>,
-    get_tsk: Box<dyn Fn(ChainEpoch) -> anyhow::Result<crate::blocks::TipsetKeys>>,
-    db: DB,
+    chain_store: Arc<ChainStore<DB>>,
     chain_config: Arc<ChainConfig>,
     bail: AtomicBool,
 }
 
-impl<DB: Blockstore> ForestExterns<DB> {
+impl<DB: Blockstore + Send + Sync + 'static> ForestExterns<DB> {
     pub fn new(
         rand: impl Rand + 'static,
+        heaviest_tipset: Arc<Tipset>,
         epoch: ChainEpoch,
         root: Cid,
-        lookback: Box<dyn Fn(ChainEpoch) -> anyhow::Result<Cid>>,
-        get_tsk: Box<dyn Fn(ChainEpoch) -> anyhow::Result<crate::blocks::TipsetKeys>>,
-        db: DB,
+        chain_store: Arc<ChainStore<DB>>,
         chain_config: Arc<ChainConfig>,
     ) -> Self {
         ForestExterns {
             rand: Box::new(rand),
+            heaviest_tipset,
             epoch,
             root,
-            lookback,
-            get_tsk,
-            db,
+            chain_store,
             chain_config,
             bail: AtomicBool::new(false),
         }
+    }
+
+    fn get_lookback_tipset_state_root_for_round(&self, height: ChainEpoch) -> anyhow::Result<Cid> {
+        let (_, st) = self.chain_store.get_lookback_tipset_for_round(
+            Arc::clone(&self.chain_config),
+            Arc::clone(&self.heaviest_tipset),
+            height,
+        )?;
+        Ok(st)
     }
 
     fn worker_key_at_lookback(
@@ -76,20 +84,20 @@ impl<DB: Blockstore> ForestExterns<DB> {
             );
         }
 
-        let prev_root = (self.lookback)(height)?;
-        let lb_state = StateTree::new_from_root(&self.db, &prev_root)?;
+        let prev_root = self.get_lookback_tipset_state_root_for_round(height)?;
+        let lb_state = StateTree::new_from_root(&self.chain_store.db, &prev_root)?;
 
         let actor = lb_state
             .get_actor(miner_addr)?
             .ok_or_else(|| anyhow::anyhow!("actor not found {:?}", miner_addr))?;
 
-        let tbs = TrackingBlockstore::new(&self.db);
+        let tbs = TrackingBlockstore::new(&self.chain_store.db);
 
         let ms = fil_actor_interface::miner::State::load(&tbs, actor.code, actor.state)?;
 
         let worker = ms.info(&tbs)?.worker.into();
 
-        let state = StateTree::new_from_root(&self.db, &self.root)?;
+        let state = StateTree::new_from_root(&self.chain_store.db, &self.root)?;
 
         let addr = resolve_to_key_addr(&state, &tbs, &worker)?;
 
@@ -117,11 +125,13 @@ impl<DB: Blockstore> ForestExterns<DB> {
     }
 }
 
-impl<DB: Blockstore> Externs for ForestExterns<DB> {}
+impl<DB: Blockstore + Send + Sync + 'static> Externs for ForestExterns<DB> {}
 
-impl<DB> Chain for ForestExterns<DB> {
+impl<DB: Blockstore> Chain for ForestExterns<DB> {
     fn get_tipset_cid(&self, epoch: ChainEpoch) -> anyhow::Result<Cid> {
-        (self.get_tsk)(epoch)?.cid()
+        self.chain_store
+            .get_epoch_tsk(Arc::clone(&self.heaviest_tipset), epoch)?
+            .cid()
     }
 }
 
@@ -145,7 +155,7 @@ impl<DB> Rand for ForestExterns<DB> {
     }
 }
 
-impl<DB: Blockstore> Consensus for ForestExterns<DB> {
+impl<DB: Blockstore + Send + Sync + 'static> Consensus for ForestExterns<DB> {
     // See https://github.com/filecoin-project/lotus/blob/v1.18.0/chain/vm/fvm.go#L102-L216 for reference implementation
     fn verify_consensus_fault(
         &self,